Refactor extract processAll function into its own file
