+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules
+coverage
 test
 layout_src
 .env
