+++ macros/phoscur.ts
@@ -1,5 +1,4 @@
 
-
 export function prepare(newMacro) {
 
   const macroExtensions = {
@@ -7,6 +6,7 @@ export function prepare(newMacro) {
         .tapKey("X_ENTER")
         .typeAlphanumeric("(gg)")
         .tapKey("X_ENTER"),
-  }
-  return { macroExtensions }
+  };
+
+  return { macroExtensions };
 }
